Revert "Merge branch 'Development-v2' of https://github.com/muinelee/Mineral_Mayhem into Match-End-Fixes-RC"
This reverts commit 51b1c06491591eb222bc2b63150712cfc0f1e1af, reversing
changes made to 28fea8ececf31eabd35c51e49d58ba22482f20ee.

+++ Assets/Scripts/Core/CoreBehaviour.cs
@@ -36,7 +36,7 @@ public class CoreBehaviour : NetworkBehaviour, IHealthComponent
         team = NetworkPlayer.Team.Neutral;
     }
 
-    public void OnTakeDamage(float damageAmount)
+    public void OnTakeDamage(int damageAmount)
     {
         if (!Object.HasStateAuthority) return;
